requirements.txt
# Core dependencies for EEG data collection and analysis
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
pandas>=1.3.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# Web dashboard and visualization
streamlit>=1.28.0
plotly>=5.17.0
pillow>=8.0.0
opencv-python>=4.5.0

# EEG data processing
mne>=1.0.0
mne-connectivity>=0.3.0

# Real-time streaming
pylsl>=1.16.0
pyserial>=3.5

# EEG headset specific libraries
muselsl>=2.1.0
pyOpenBCI>=0.13
thinkgear>=0.1.2

# Signal processing
pywavelets>=1.4.0
antropy>=0.1.4

# Machine learning
tensorflow>=2.10.0
torch>=1.12.0

# Utilities
psutil>=5.9.0
threading-timer>=0.3.0
asyncio>=3.4.3
websockets>=10.0

# Audio feedback
pygame>=2.1.0
pydub>=0.25.0

# Data storage
h5py>=3.7.0
sqlite3
